채팅 기능 추가, 제품 상세 페이지 디자인 수정

+++ jsp_project/src/main/java/com/carrot/dao/ProductDAO.java
@@ -122,11 +122,11 @@ public class ProductDAO extends BaseDAO{
 	}
 	
 	// ID로 상품 조회
-	public ProductDTO selectProductById(long id) {
+	public ProductDTO selectProductById(int id) {
 		String sql = "SELECT * FROM PRODUCT WHERE PRODUCT_ID = ?";
 
 		try (Connection conn = getConnection(); PreparedStatement pstmt = conn.prepareStatement(sql)) {
-			pstmt.setLong(1, id);
+			pstmt.setInt(1, id);
 			try (ResultSet rs = pstmt.executeQuery()) {
 				if (rs.next()) {
 					return ProductDTO.builder()
